change in labour table

+++ database/migrations/2024_03_10_085414_create_labour.php
@@ -13,12 +13,12 @@ return new class extends Migration
     {
         Schema::create('labour', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('gender_id');
-            $table->unsignedBigInteger('district_id');
-            $table->unsignedBigInteger('taluka_id');
-            $table->unsignedBigInteger('village_id');
-            $table->unsignedBigInteger('skill_id');
+            $table->integer('user_id');
+            $table->integer('gender_id');
+            $table->integer('district_id');
+            $table->integer('taluka_id');
+            $table->integer('village_id');
+            $table->integer('skill_id');
             $table->string('user_type')->default('null');
             $table->string('full_name');
             $table->string('date_of_birth');
@@ -29,7 +29,7 @@ return new class extends Migration
             $table->string('longitude');
             $table->integer('is_approved')->default(true);
             $table->integer('is_resubmitted')->default(false);
-            $table->unsignedBigInteger('reason_id')->nullable();
+            $table->integer('reason_id')->nullable();
             $table->string('other_remark')->default('null');
             $table->string('sync_reason')->default('null');
             $table->string('aadhar_image')->default('null');
